requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
yt-dlp
ffmpeg-python